Replaced almost all Hashtable and NameValueCollection uses to Dictionary<string,string>, and ArrayList to List<string>.
This not only uses more .NET-style strong-type coding, but gets rid of some uncertain type casts as well.

Dictionary<string,string> is superior to NameValueCollection because the newer class has more support for it in .NET going forward, and more deterministic value getters.  For example, NVC is ambiguous when it comes to getting a key that does not exist vs. a key whose value is null, where as a Dictionary<TKey,TValue> forces you to deal with the difference.